Connected to twitter api
- renamed the directory of python-app to backend
- moved tests into tests/ directory
= connected to twitter api
