+++ .gitignore
@@ -1,4 +1,5 @@
 .DS_Store
 *.pyc
-*/node_modules/
+node_modules/
+instance/
 frontend/public/bundle.js
