+++ backend/config.py
@@ -1,0 +1,4 @@
+CONSUMER_KEY = 'not really this'
+CONSUMER_SECRET = 'not really this'
+ACCESS_TOKEN = 'not really this'
+ACCESS_SECRET = 'not really this'
